+++ lovely/essequence/__init__.py
@@ -1,2 +1,2 @@
-VERSION = '0.0.2'
+VERSION = '0.1.0'
 from lovely.essequence.sequence import Sequence  # noqa