+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changes for lovely.essequence
 unreleased
 ==========
 
+2016/07/19 0.1.0
+================
+
+Note:
+    This version is not backward compatible.
+
 - made the sequence index crate compatible
 - renamed sequence index